// frontend/src/app/models/customer.model.ts
export interface Customer {
  customerId: number;
  customerName: string;
  streetNumber?: string;
  streetName?: string;
  unitNumber?: string;
  city?: string;
  state?: string;
  postCode?: string;
  emailId?: string;
  mobileNumber?: string;
  communicationPreference?: string;
  serial?: string;
  createdAt?: string;
  updatedAt?: string;
}